fix(document): harden decoding and extreme layout
Preserve strict Unicode decoding across BOM selection and make document flow
safe for non-breaking separators, saturated embedded widths, insecure NUL
input, and disposal boundaries.

+++ src/SharpVision.Document/Controls/Documents/DocumentLinkRegion.cs
@@ -45,5 +45,5 @@ internal readonly struct DocumentLinkRegion
     /// <param name="column">The candidate column.</param>
     /// <returns>True when the column is covered.</returns>
     [Pure]
-    public bool Contains(int column) => column >= Column && column < Column + Cells;
+    public bool Contains(int column) => column >= Column && column < (long) Column + Cells;
 }
